+++ wiliwili/include/api/bilibili/result/home_hots_history_result.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "nlohmann/json.hpp"
+#include "bilibili/util/json.hpp"
 #include "user_result.h"
 #include "home_result.h"
 
@@ -12,9 +12,9 @@ namespace bilibili {
 
 class HotsHistoryVideoResult {
 public:
-    int aid;
+    uint64_t aid;
     std::string bvid;
-    int cid;
+    uint64_t cid;
     std::string pic;
     std::string title;
     int duration;
